Refactor parameter loading to use YAML file
